+++ lib/examples.rb
@@ -5,5 +5,7 @@ require 'active_support/all'
 # require 'examples/my_dice'
 # require 'examples/my_dynamic_dice'
 require 'marskal/module_configurator'
+require 'marskal/core_ext'
 require 'examples/dice_method_one'
+require 'examples/dice_method_two'